LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/getter-special-cases/csp-sandbox-yes.https-expected.txt


FAIL originAgentCluster must equal true assert_equals: expected (boolean) true but got (undefined) undefined
